refactor(core): use getters and readonly arrays
